+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Introduced a `GistFS` tab, to allow you choose between managing Gists in it's own side-bar, or via the `Gists` view on the `Explorer` tab
 * Added the `Change Description` command to the `Gists` view, which allows changing an existing Gist's description
+* Added the `Rename File` command to the `Gists` view, which allows renaming an existing file in a Gist
 * Added the `Add Active File` command to the `Gists` view, which allows "uploading" the active editor to the specified Gist
  
 ## v0.0.5 (11/27/2016)